Extracting configs.
